requirements.txt
torch>=1.5.0,<1.6.0
torchvision>=0.6.0,<0.7.0
numpy
sklearn
torchnet
pillow
umap-learn
pandas
matplotlib
tensorboard
tensorboard_logger